backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pydantic
python-multipart
pytesseract
pillow
pypdfium2
python-dotenv
sqlite-utils
google-generativeai
